+++ android/src/main/java/dev/barooni/capacitor/calendar/CapacitorCalendarPlugin.kt
@@ -19,6 +19,7 @@ import dev.barooni.capacitor.calendar.models.inputs.DeleteCalendarInput
 import dev.barooni.capacitor.calendar.models.inputs.DeleteEventInput
 import dev.barooni.capacitor.calendar.models.inputs.DeleteEventWithPromptInput
 import dev.barooni.capacitor.calendar.models.inputs.DeleteEventsByIdInput
+import dev.barooni.capacitor.calendar.models.inputs.GetDefaultCalendarInput
 import dev.barooni.capacitor.calendar.models.inputs.ListEventsInRangeInput
 import dev.barooni.capacitor.calendar.models.inputs.ModifyCalendarInput
 import dev.barooni.capacitor.calendar.models.inputs.ModifyEvent
@@ -30,6 +31,7 @@ import dev.barooni.capacitor.calendar.models.results.CreateEventWithPromptResult
 import dev.barooni.capacitor.calendar.models.results.ModifyEventWithPromptResult
 import dev.barooni.capacitor.calendar.models.results.RequestAllPermissionsResult
 import dev.barooni.capacitor.calendar.models.results.RequestPermissionResult
+import dev.barooni.capacitor.calendar.models.templates.JSResult
 
 @CapacitorPlugin(
     name = "CapacitorCalendar",
@@ -307,10 +309,16 @@ class CapacitorCalendarPlugin : Plugin() {
     @PluginMethod(returnType = PluginMethod.RETURN_PROMISE)
     fun getDefaultCalendar(call: PluginCall) {
         try {
-            val result = implementation.getDefaultCalendar()
-            call.resolve(result.toJSON())
+            val input = GetDefaultCalendarInput(call)
+            implementation.getDefaultCalendar(input) { result, error ->
+                if (error != null) {
+                    rejectCall(call, error)
+                    return@getDefaultCalendar
+                }
+                resolveCall(call, result)
+            }
         } catch (error: Exception) {
-            call.reject(error.message)
+            rejectCall(call, error)
         }
     }
 
@@ -454,5 +462,27 @@ class CapacitorCalendarPlugin : Plugin() {
     @PluginMethod
     fun updateRemindersList(call: PluginCall) {
         call.unimplemented(PluginError.Unimplemented(::updateRemindersList.name).message)
+    }
+
+    private fun rejectCall(
+        call: PluginCall,
+        error: Exception?,
+    ) {
+        call.reject(error?.message ?: PluginError.CustomError("An unknown error has occurred.").message)
+    }
+
+    private fun resolveCall(call: PluginCall) {
+        resolveCall(call, null)
+    }
+
+    private fun resolveCall(
+        call: PluginCall,
+        result: JSResult?,
+    ) {
+        if (result != null) {
+            call.resolve(result.toJSON())
+        } else {
+            call.resolve()
+        }
     }
 }