+++ android/src/main/java/dev/barooni/capacitor/calendar/models/inputs/GetDefaultCalendarInput.kt
@@ -1,0 +1,9 @@
+package dev.barooni.capacitor.calendar.models.inputs
+
+import com.getcapacitor.PluginCall
+
+data class GetDefaultCalendarInput(
+    private val call: PluginCall,
+) {
+    val useFallbackCalendar: Boolean = call.getBoolean("useFallbackCalendar", false)
+}