+++ ios/Plugin/Models/Inputs/GetDefaultCalendarInput.swift
@@ -1,0 +1,13 @@
+import Capacitor
+
+struct GetDefaultCalendarInput {
+    private let useFallbackCalendar: Bool
+
+    init(call: CAPPluginCall) {
+        self.useFallbackCalendar = call.getBool("useFallbackCalendar", false)
+    }
+
+    func getUseFallbackCalendar() -> Bool {
+        return useFallbackCalendar
+    }
+}